新接口下增加vtk模块

+++ fealpy/experimental/backend/pytorch_backend.py
@@ -457,7 +457,7 @@ class PyTorchBackend(Backend[Tensor], backend_name='pytorch'):
             mi = cls.multi_index_matrix(p, TD, dtype=torch.int)
 
         c = torch.arange(1, p+1, dtype=itype, device=device)
-        P = 1.0 / torch.cumprod(c, dim=0)
+        P = 1.0 / torch.cumprod(c, dim=0, dtype=bc.dtype)
         t = torch.arange(0, p, dtype=itype, device=device)
         Ap = p*bc.unsqueeze(-2) - t.reshape(-1, 1)
         Ap = torch.cumprod(Ap, dim=-2).clone()
